feat: 移除 Plan Mode——规划成为 Agent 原生能力，全自动无确认
